Move KeyValError into tokeniser module
