+++ srctools/tokenizer.py
@@ -1,0 +1,46 @@
+from typing import (
+    Union, Optional,
+    Callable, Iterable,
+)
+
+
+class TokenSyntaxError(Exception):
+    """An error that occurred when parsing a file.
+
+    mess = The error message that occurred.
+    file = The filename passed to Property.parse(), if it exists
+    line_num = The line where the error occurred.
+    """
+    def __init__(
+            self,
+            message: str,
+            file: Optional[str],
+            line: Optional[int]
+            ) -> None:
+        super().__init__()
+        self.mess = message
+        self.file = file
+        self.line_num = line
+
+    def __repr__(self):
+        return 'ParseError({!r}, {!r}, {!r})'.format(
+            self.mess,
+            self.file,
+            self.line_num,
+            )
+
+    def __str__(self):
+        """Generate the complete error message.
+
+        This includes the line number and file, if available.
+        """
+        mess = self.mess
+        if self.line_num:
+            mess += '\nError occurred on line ' + str(self.line_num)
+            if self.file:
+                mess += ', with file'
+        if self.file:
+            if not self.line_num:
+                mess += '\nError occurred with file'
+            mess += ' "' + self.file + '"'
+        return mess
